requirements.txt
fastapi
uvicorn
firebase-admin
anthropic
python-multipart
openai
google-genai
